[Feature] 포트리스 모작 #2

+++ WindowsAPI/Constants.h
@@ -4,4 +4,4 @@ constexpr int32 GWinSizeX = 800;
 constexpr int32 GWinSizeY = 600;
 constexpr int32 GMinimapSizeX = 200;
 constexpr int32 GMinimapSizeY = 128;
-const double PI = ::atan(1) * 4;
+const float PI = ::atan(1) * 4;
